move tags into their own modules
