+++ src/tags/mod.rs
@@ -1,8 +1,6 @@
 //! [4.3. Playlist Tags]
 //!
 //! [4.3. Playlist Tags]: https://tools.ietf.org/html/rfc8216#section-4.3
-use crate::{ErrorKind, Result};
-use trackable::error::ErrorKindExt;
 
 macro_rules! may_invalid {
     ($expr:expr) => {
@@ -20,24 +18,17 @@ macro_rules! impl_from {
     };
 }
 
-pub use self::basic::{ExtM3u, ExtXVersion};
-pub use self::master_playlist::{
-    ExtXIFrameStreamInf, ExtXMedia, ExtXSessionData, ExtXSessionKey, ExtXStreamInf,
-};
-pub use self::media_or_master_playlist::{ExtXIndependentSegments, ExtXStart};
-pub use self::media_playlist::{
-    ExtXDiscontinuitySequence, ExtXEndList, ExtXIFramesOnly, ExtXMediaSequence, ExtXPlaylistType,
-    ExtXTargetDuration,
-};
-pub use self::media_segment::{
-    ExtInf, ExtXByteRange, ExtXDateRange, ExtXDiscontinuity, ExtXKey, ExtXMap, ExtXProgramDateTime,
-};
-
 mod basic;
 mod master_playlist;
-mod media_or_master_playlist;
 mod media_playlist;
 mod media_segment;
+mod shared;
+
+pub use basic::*;
+pub use master_playlist::*;
+pub use media_playlist::*;
+pub use media_segment::*;
+pub use shared::*;
 
 /// [4.3.4. Master Playlist Tags]
 ///
@@ -114,16 +105,3 @@ impl_from!(MediaSegmentTag, ExtXDiscontinuity);
 impl_from!(MediaSegmentTag, ExtXKey);
 impl_from!(MediaSegmentTag, ExtXMap);
 impl_from!(MediaSegmentTag, ExtXProgramDateTime);
-
-fn parse_yes_or_no(s: &str) -> Result<bool> {
-    match s {
-        "YES" => Ok(true),
-        "NO" => Ok(false),
-        _ => track_panic!(ErrorKind::InvalidInput, "Unexpected value: {:?}", s),
-    }
-}
-
-fn parse_u64(s: &str) -> Result<u64> {
-    let n = track!(s.parse().map_err(|e| ErrorKind::InvalidInput.cause(e)))?;
-    Ok(n)
-}
