+++ src/tags/shared/start.rs
@@ -1,35 +1,9 @@
-use super::parse_yes_or_no;
+use crate::utils::parse_yes_or_no;
 use crate::attribute::AttributePairs;
 use crate::types::{ProtocolVersion, SignedDecimalFloatingPoint};
 use crate::{Error, ErrorKind, Result};
 use std::fmt;
 use std::str::FromStr;
-
-/// [4.3.5.1. EXT-X-INDEPENDENT-SEGMENTS]
-///
-/// [4.3.5.1. EXT-X-INDEPENDENT-SEGMENTS]: https://tools.ietf.org/html/rfc8216#section-4.3.5.1
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct ExtXIndependentSegments;
-impl ExtXIndependentSegments {
-    pub(crate) const PREFIX: &'static str = "#EXT-X-INDEPENDENT-SEGMENTS";
-
-    /// Returns the protocol compatibility version that this tag requires.
-    pub fn requires_version(self) -> ProtocolVersion {
-        ProtocolVersion::V1
-    }
-}
-impl fmt::Display for ExtXIndependentSegments {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        Self::PREFIX.fmt(f)
-    }
-}
-impl FromStr for ExtXIndependentSegments {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self> {
-        track_assert_eq!(s, Self::PREFIX, ErrorKind::InvalidInput);
-        Ok(ExtXIndependentSegments)
-    }
-}
 
 /// [4.3.5.2. EXT-X-START]
 ///
@@ -39,6 +13,7 @@ pub struct ExtXStart {
     time_offset: SignedDecimalFloatingPoint,
     precise: bool,
 }
+
 impl ExtXStart {
     pub(crate) const PREFIX: &'static str = "#EXT-X-START:";
 
@@ -74,6 +49,7 @@ impl ExtXStart {
         ProtocolVersion::V1
     }
 }
+
 impl fmt::Display for ExtXStart {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", Self::PREFIX)?;
@@ -84,6 +60,7 @@ impl fmt::Display for ExtXStart {
         Ok(())
     }
 }
+
 impl FromStr for ExtXStart {
     type Err = Error;
     fn from_str(s: &str) -> Result<Self> {
@@ -115,15 +92,6 @@ impl FromStr for ExtXStart {
 #[cfg(test)]
 mod test {
     use super::*;
-
-    #[test]
-    fn ext_x_independent_segments() {
-        let tag = ExtXIndependentSegments;
-        let text = "#EXT-X-INDEPENDENT-SEGMENTS";
-        assert_eq!(text.parse().ok(), Some(tag));
-        assert_eq!(tag.to_string(), text);
-        assert_eq!(tag.requires_version(), ProtocolVersion::V1);
-    }
 
     #[test]
     fn ext_x_start() {
